requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
supabase==2.0.3
python-engineio==4.8.0
python-socketio==5.10.0
werkzeug==3.0.1